create useLocation hook and fix PR issues
